Set up architecture for descent parser
This required a lot of knowledge scraped from the Python Language
Reference, which more-or-less describes a context-free grammar for
all of Python. I cherrypicked the necessary symbols and made structs
for each of them. In particular, the plan is to support:

    * Literal values: ints, strings, lists, dicts
    * Variables of the above literals, supporting multiple assignment
    * List/dict member access and assignment
    * A del command

Additionally, I believe the parser will also support various types
of arithmetical operations and attribute accessors.

The other bits of architecture, like the read-eval-print-loop, are
just placeholders for when the major heavy lifting will occur.