Add proof-of-concept for parsing statements
Specifically, the program will now detect when input statements
begin with the "del" keyword! Nifty!

The code, however, is ugly. I just wanted to make sure I wasn't
absolutely insane in how I set it up so far. Most of it will be
rewritten immediately.

The idea is to quickly determine whether a given statement is
a deletion statement, an assignment statement of some sort, or
something else (i.e. malformed). Then, the statement will be built
up from scratch in the main loop. Thinking about it now, there will
have to be quite a bit of bookkeeping, so it may have to change a
bit... but that's a plan for another day.

I also added some include guards.