Enable reading lists of targets
Note that this code will not compile. There are too many functions
that need to be written that have not been written yet. However,
the read_target_list() function should be done, more or less.

I'm jumping the gun by writing parsing before tokenizing. Mostly,
I want to write as much of the parser as I can so that I know what
the tokenizer is supposed to look like (which tokens should exist,
etc.).