Add most read_*() functions to parser
Here's the scoop: this is really bad code. I am going to fix it,
and soon. But first I want to verify that I am not completely
insane in thinking this will work, so for now I am not letting
perfection be the enemy of completion. A few parts are more
egregious than others, but overall the code is just too long.

Enough of that. The cascade should be complete in the sense that
del-statements should be completely parseable, now. (I haven't
yet added make_() functions, but that's easy and will happen
right away.) It should even compile and run soon.

A tentative TODO list (mostly for myself, but hey, if you're
reading this, good for you):
 * Write the make_*() functions
 * Write next_token() completely
 * Verify that the code works for del-statements
 * Make everything less terrible
 * Support assignment statements, expression lists, etc
 * Verify again that the code works
 * Move on to evaluation

After evaluation (which is about as big a beast as parsing, though
maybe a bit less repetitive), the interpreter should be done! Then,
people can add garbage collection to it... :)