Write all make_*() allocation functions
Had to refactor some previous code, as well. I discovered that it
is going to be quite a bit more difficult to distinguish between
idents and strings, because one is a subset of the other, but I
hope I can bank on the fact that if you're looking for an ident,
but get a string of some other sort, you have no backup and never
ask to see any other tokens afterwards. I'm scared of this
assumption, though. I don't know what the future holds.