Fix identifier and literal tokenizing/parsing
The aforementioned problem, i.e. that I was treating single and
double-quotes as tokens, has been resolved. Stringy things should
be tokenized and parsed properly now.

With that, the code revolving around del-statements should,
technically, be done. However, it segfaults immediately (as
expected!! I'm doing nothing with the pointers I'm generating).
Thus, what remains is to verify that it's working by implementing
lots of printer functions. Also, I have to fix a lot of the code
because it's garbage.